Add color, thickness and alpha extraction
this is very beta. The exporter will (if specified in the UI) try
to extract color, and will assign correct thickness and alpha values to
gpencil strokes.

create colors are cleaned up when no longer needed + only created when
no other color with the same color (hex-based, so a little lossy to
solve float comparison problems) is present.